Move Schema from core/GraphQLIRVisitor to core/Schema
