+++ types/relay-compiler/lib/language/RelayLanguagePluginInterface.d.ts
@@ -2,7 +2,7 @@ import { Root, Fragment, GeneratedDefinition } from '../core/GraphQLIR';
 import { IRTransform } from '../core/GraphQLCompilerContext';
 import { GeneratedNode, RelayConcreteNode } from 'relay-runtime';
 import { ScalarTypeMapping } from './javascript/RelayFlowTypeTransformers';
-import { Schema } from '../core/GraphQLIRVisitor';
+import { Schema } from '../core/Schema';
 
 /**
  * A language plugin allows relay-compiler to both read and write files for any
